Disables project caching in CI
Something's up with trying to restore from the local project cache.

It appears as if GHC is rebuilding some library dependencies in one
build step, but attempting to restore a version of the local project
cache that was compiled against an earlier version of that library.

This probably means that the cache restore key should be a bit more
granular (perhaps it should depend on the freeze file hash, which would
mean that older snapshots of the local cache would never restore in the
presence of a newer Hackage snapshot).